nodeutils: Rename back to netconf package, better function naming
