+++ nodes/c8000/c8000.go
@@ -13,8 +13,8 @@ import (
 	"regexp"
 
 	"github.com/charmbracelet/log"
+	"github.com/srl-labs/containerlab/netconf"
 	"github.com/srl-labs/containerlab/nodes"
-	"github.com/srl-labs/containerlab/nodeutils"
 	"github.com/srl-labs/containerlab/types"
 	"github.com/srl-labs/containerlab/utils"
 )
@@ -79,7 +79,7 @@ func (n *c8000) PreDeploy(ctx context.Context, params *nodes.PreDeployParams) er
 }
 
 func (n *c8000) SaveConfig(_ context.Context) error {
-	err := nodeutils.NetconfSaveConfig(n.Cfg.LongName,
+	err := netconf.SaveRunningConfig(n.Cfg.LongName,
 		defaultCredentials.GetUsername(),
 		defaultCredentials.GetPassword(),
 		scrapliPlatformName,
